requirements.txt
streamlit>=1.20.0
openai>=0.27.0
python-dotenv
python-pptx
